requirements.txt
pygame==2.5.2
gspread==5.12.0
google-auth==2.23.4
numpy==1.24.4